Refactor ticket page
